requirements.txt
clipman
pillow
requests
pyautogui
pyscreenshot
